[Sprint 56][S56-004] Split start handler into focused view slices
Extract notification settings rendering and seller auction presentation helpers from the oversized start handler into dedicated modules while preserving router behavior and callback contracts. Keep existing tests and imports stable by re-exporting helper symbols through start.py.
